requirements.txt
inquirer
colorama
pyperclip
